refactor: Result 클래스 분리
