+++ src/main/java/baseball/Computer.java
@@ -1,0 +1,4 @@
+package baseball;
+
+public class Computer {
+}
